requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0

# Google Cloud dependencies
google-cloud-vision>=3.4.5
google-cloud-storage>=2.13.0
google-generativeai>=0.7.0  # Updated for Gemini 2.0 support

# PDF processing
PyMuPDF>=1.23.7  # For PDF processing (fitz)

# HTTP client for URL fetching
httpx>=0.25.0  # For fetching PDFs from URLs

# Database
# SQLite is built-in to Python

# Utilities
tqdm>=4.66.1  # For progress bars
tenacity>=8.2.3  # For retries
uuid>=1.30  # For generating unique IDs 